+++ RELEASE_NOTES.md
@@ -18,7 +18,7 @@ Notable changes include:
         * CPU & GPU unit testing of the public interface.
       * RankNTensor (Third, Fourth, Fifth) have been refactored to execute on the GPU.
       * Spheral::NodePairListView allows for implicit data migration of Spheral::NodePairList data.
-      * Created view classes for the Quadratic and CubicHermite interpolator classes, called QIView and CHIView, respectively.
+      * Created view classes for the Quadratic and CubicHermite interpolator classes.
         * These follow the inheritance pattern used previously.
       * Optimizations to RankTensor types:
         * Stack allocation of tensor data; Static casting for CRTP implementation.